Changed units_convert type

+++ src/rappture_io.f90
@@ -2999,11 +2999,11 @@ character*100  function rp_lib_get_wrap( driver, path)
   call rp_lib_get( driver, path, rp_lib_get_wrap)
 end function rp_lib_get_wrap
 
-integer function readGenericDbl( status, element_to_read, short_name, InputEcho, units_convert)
+real*8 function readGenericDbl( status, element_to_read, short_name, InputEcho, units_convert)
   integer, intent(inout) :: status
   character(len=*), intent(in) :: element_to_read, short_name
   character*1500, intent(inout) :: InputEcho
-  real*8, intent(in) :: units_convert
+  integer, intent(in) :: units_convert
 
   character*100 :: strVal
   integer :: rp_units_convert_dbl
